+++ src/front/asm.cpp
@@ -42,7 +42,7 @@ bool EXPAND_ONLY = false;
 bool EARLY_VERIFICATION_ONLY = false;
 // are we only checking what size will the bytecode by?
 bool REPORT_BYTECODE_SIZE = false;
-bool VERIFY_REGISTER_ACCESS = false;
+bool PERFORM_STATIC_ANALYSIS = true;
 
 bool VERBOSE = false;
 bool DEBUG = false;
@@ -89,7 +89,7 @@ static bool usage(const char* program, bool show_help, bool show_version, bool v
              << "    " << "-C, --verify             - verify source code correctness without actually compiling it\n"
              << "    " << "                           this option turns assembler into source level debugger and static code analyzer hybrid\n"
              << "    " << "    --size               - calculate and report final bytecode size\n"
-             << "    " << "    --static-check       - enable static checking (alpha)\n"
+             << "    " << "    --no-sa              - disable static checking of register accesses (use in case of false positives)\n"
              ;
     }
 
@@ -184,8 +184,8 @@ int main(int argc, char* argv[]) {
         } else if (option == "--size") {
             REPORT_BYTECODE_SIZE = true;
             continue;
-        } else if (option == "--static-check") {
-            VERIFY_REGISTER_ACCESS = true;
+        } else if (option == "--no-sa") {
+            PERFORM_STATIC_ANALYSIS = false;
             continue;
         } else if (str::startswith(option, "-")) {
             cout << "error: unknown option: " << option << endl;
@@ -306,7 +306,7 @@ int main(int argc, char* argv[]) {
         assembler::verify::framesHaveOperands(expanded_lines);
         assembler::verify::framesHaveNoGaps(expanded_lines, expanded_lines_to_source_lines);
         assembler::verify::blocksEndWithFinishingInstruction(expanded_lines);
-        if (VERIFY_REGISTER_ACCESS) {
+        if (PERFORM_STATIC_ANALYSIS) {
             assembler::verify::manipulationOfDefinedRegisters(cooked_tokens, DEBUG);
         }
     } catch (const pair<unsigned, string>& e) {